+++ pkg/participant/nn_party.go
@@ -2,6 +2,9 @@ package participant
 
 import (
 	"fmt"
+	"math"
+	"math/rand"
+	"time"
 
 	"gonum.org/v1/gonum/mat"
 
@@ -93,14 +96,39 @@ func NewNNParty(party *Party) *NNParty {
 // inputSize: 输入层神经元数量
 // outputSize: 输出层神经元数量
 func (nnp *NNParty) InitializeLayer(inputSize int, outputSize int, activation NNActivationFunc, initMethod WeightInitMethod) {
+	// 设置随机种子，确保权重初始化的可重现性
+	rand.Seed(time.Now().UnixNano() + int64(nnp.ID))
+
 	// 根据激活函数类型和初始化方法两个枚举类设置对应的权重矩阵初始化
-	//这里仅作测试示例 后期需要重新写权重初始化
 	weights := mat.NewDense(outputSize, inputSize, nil)
-	for i := 0; i < outputSize; i++ {
-		for j := 0; j < inputSize; j++ {
-			// 简单的权重初始化策略（实际应用中可能需要更复杂的初始化方法）
-			weights.Set(i, j, 0.1*(float64(i+j)/float64(inputSize+outputSize)))
+	// 根据初始化方法选择不同的策略
+	switch initMethod {
+	case InitHe:
+		// He初始化：适用于ReLU激活函数
+		scale := math.Sqrt(2.0 / float64(inputSize))
+		for i := 0; i < outputSize; i++ {
+			for j := 0; j < inputSize; j++ {
+				weights.Set(i, j, rand.NormFloat64()*scale)
+			}
 		}
+	case InitXavier:
+		// Xavier初始化：适用于Sigmoid/Tanh激活函数
+		scale := math.Sqrt(2.0 / float64(inputSize+outputSize))
+		for i := 0; i < outputSize; i++ {
+			for j := 0; j < inputSize; j++ {
+				weights.Set(i, j, rand.NormFloat64()*scale)
+			}
+		}
+	case InitRandom:
+		// 简单随机初始化
+		for i := 0; i < outputSize; i++ {
+			for j := 0; j < inputSize; j++ {
+				weights.Set(i, j, rand.NormFloat64()*0.1)
+			}
+		}
+	case InitZero:
+		// 零初始化（权重全为0）
+		// 权重矩阵默认为0，无需额外设置
 	}
 
 	// 将偏置初始化为零向量
@@ -148,4 +176,83 @@ func (nnp *NNParty) GetLayerInfo() string {
 
 	return fmt.Sprintf("参与方 %d: 层维度 %dx%d，激活函数: %v",
 		nnp.ID, nnp.LayerData.InputSize, nnp.LayerData.OutputSize, nnp.LayerData.Activation)
+}
+
+// 激活函数实现（用于同态加密的多项式近似）
+
+// ReLUPolyApprox ReLU激活函数的多项式近似
+// 使用Chebyshev多项式近似ReLU函数
+func ReLUPolyApprox(x float64) float64 {
+	// 简单的多项式近似：max(0, x) ≈ 0.5 * (x + sqrt(x^2 + ε))
+	// 其中ε是一个小的正数，避免除零
+	epsilon := 1e-6
+	if x >= 0 {
+		return x
+	}
+	return 0.5 * (x + math.Sqrt(x*x+epsilon))
+}
+
+// SigmoidPolyApprox Sigmoid激活函数的多项式近似
+// 使用Chebyshev多项式近似Sigmoid函数
+func SigmoidPolyApprox(x float64) float64 {
+	// 简单的多项式近似：1/(1 + e^(-x)) ≈ 0.5 * (1 + x/sqrt(1 + x^2))
+	// 这个近似在[-3, 3]范围内比较准确
+	if x > 3 {
+		return 1.0
+	}
+	if x < -3 {
+		return 0.0
+	}
+	return 0.5 * (1.0 + x/math.Sqrt(1.0+x*x))
+}
+
+// SoftmaxPolyApprox Softmax激活函数的多项式近似
+// 注意：Softmax通常需要整个向量，这里提供单个元素的近似
+func SoftmaxPolyApprox(x float64, maxVal float64) float64 {
+	// 简化的softmax近似：exp(x - max) / sum(exp(x_i - max))
+	// 这里只处理单个元素，实际使用时需要整个向量
+	expVal := math.Exp(x - maxVal)
+	return expVal
+}
+
+// GetActivationFunction 根据激活函数类型返回对应的函数
+func (nnp *NNParty) GetActivationFunction() func(float64) float64 {
+	switch nnp.LayerData.Activation {
+	case ActivationReLU:
+		return ReLUPolyApprox
+	case ActivationSigmoid:
+		return SigmoidPolyApprox
+	case ActivationSoftmax:
+		// Softmax需要特殊处理，返回一个默认函数
+		return func(x float64) float64 { return x }
+	case ActivationNone:
+		return func(x float64) float64 { return x }
+	default:
+		return func(x float64) float64 { return x }
+	}
+}
+
+// ValidateLayerConfig 验证层配置的合理性
+func (nnp *NNParty) ValidateLayerConfig() error {
+	if nnp.LayerData == nil {
+		return fmt.Errorf("参与方 %d: 层数据未初始化", nnp.ID)
+	}
+
+	if nnp.LayerData.InputSize <= 0 || nnp.LayerData.OutputSize <= 0 {
+		return fmt.Errorf("参与方 %d: 输入或输出维度无效", nnp.ID)
+	}
+
+	if nnp.LayerData.Weights == nil || nnp.LayerData.Biases == nil {
+		return fmt.Errorf("参与方 %d: 权重或偏置未初始化", nnp.ID)
+	}
+
+	return nil
+}
+
+// GetLayerDimensions 获取层的维度信息
+func (nnp *NNParty) GetLayerDimensions() (int, int) {
+	if nnp.LayerData == nil {
+		return 0, 0
+	}
+	return nnp.LayerData.InputSize, nnp.LayerData.OutputSize
 }
